Added the information for jsoncpp
I used a library off of the internet called JsonCpp
(http://jsoncpp.sourceforge.net/) to do JSON parsing.  It was not
intended for ROS so I turned it into a ROS package.  I did not write
this library and I needed to track down the original author's name,
website, and the license this library was distributed under, which said
it was public domain.  So I've included his name in the manifest.